lfc-agent/requirements.txt
anthropic>=0.18.0
psycopg2-binary>=2.9.9
requests>=2.31.0
Pillow>=10.2.0
yt-dlp>=2026.2.4
python-dotenv>=1.0.0
schedule>=1.2.0
beautifulsoup4>=4.12.3
httpx>=0.27.0
playwright>=1.40.0
